Rename the 'extend' module in 'custom' for events
